scrapepi/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
typing_extensions==4.12.2
azure-cosmos==4.7.0
aiohttp==3.9.5
python-dateutil==2.8.2